rag-example/backend/analysis-function/requirements.txt
functions-framework==3.8.3
Flask==3.1.1
google-genai
google-auth==2.40.3
requests
